Added date ranges to popup. Only posts within range are cleaned.
Now only posts that should be cleaned will be removed from the page.
Other posts that shouldn't be cleaned (since they're unchecked in
the popup or outside of the date range) will remain on the page.

Refactored clicking/removing logic so that on the current post
is only marked and its action button is not clicked since deleting
the current post would cause us to lose our place in the Activity Log.
Instead, we constantly clean posts that have already been marked that
are not the same as the current post.